+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ## 1.7.5 - Unreleased
 
+* Fixed [#409](https://github.com/NeoVintageous/NeoVintageous/issues/409): `ga` doesn't work properly when cursor is on a NEWLINE, or TAB, or at EOF
 * Fixed [#405](https://github.com/NeoVintageous/NeoVintageous/issues/405): `v_J` should enter NORMAL mode
 
 ## 1.7.4 - 2019-01-10